#226 refactor: AuthenticationService에 인증 관련 로직 이동
